# docs/api.openapi.yml
openapi: 3.1.0
info:
  title: DispatchOps API
  version: 1.0.0
  description: Human-in-the-loop logistics control-tower API.
servers:
  - url: http://127.0.0.1:4100
paths:
  /api/health:
    get:
      summary: Check service health
      responses:
        "200": { description: Service is healthy }
  /api/auth/login:
    post:
      summary: Create an HTTP-only cookie session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200": { description: Authenticated }
        "401": { description: Invalid credentials }
  /api/overview:
    get:
      summary: Read the complete control-tower snapshot
      security: [{ cookieSession: [] }]
      responses:
        "200": { description: Operational overview }
        "401": { description: Authentication required }
  /api/agent/run:
    post:
      summary: Generate deterministic assignment proposals
      security: [{ cookieSession: [] }]
      responses:
        "200": { description: Explainable proposals generated }
        "403": { description: Dispatcher role required }
  /api/recommendations/{id}/approve:
    post:
      summary: Approve and apply one current proposal
      security: [{ cookieSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Assignment applied }
        "409": { description: Proposal is stale }
  /api/events:
    get:
      summary: Subscribe to server-sent operational events
      security: [{ cookieSession: [] }]
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: dispatchops_session